move definition of roman exceptions from plasmapy/utils/roman to plasmapy/utils/exceptions
