+++ plasmapy/utils/exceptions.py
@@ -2,7 +2,10 @@
 __all__ = [
     "PlasmaPyError",
     "PhysicsError",
+    "InvalidRomanNumeralError",
+    "OutOfRangeError",
     "RelativityError",
+    "RomanError",
     "PlasmaPyWarning",
     "CouplingWarning",
     "ImplicitUnitConversionWarning",
@@ -12,9 +15,9 @@ __all__ = [
 
 from astropy.units import UnitsWarning
 
-# ----------
-# Exceptions
-# ----------
+# ------------------------------------------------------------------------------
+#   Exceptions
+# ------------------------------------------------------------------------------
 
 
 class PlasmaPyError(Exception):
@@ -36,6 +39,14 @@ class PhysicsError(PlasmaPyError, ValueError):
     pass
 
 
+class RomanError(PlasmaPyError):
+    """A base exception for errors from `~plasmapy.utils.roman`."""
+
+    pass
+
+# ^^^^^^^^^^^^ Base Exceptions should be defined above this comment ^^^^^^^^^^^^
+
+
 class RelativityError(PhysicsError):
     """
     An exception for speeds greater than the speed of light.
@@ -44,9 +55,26 @@ class RelativityError(PhysicsError):
     pass
 
 
-# ----------
-# Warnings:
-# ----------
+class OutOfRangeError(RomanError):
+    """
+    An exception to be raised for integers that outside of the range
+    that can be converted to Roman numerals.
+    """
+
+    pass
+
+
+class InvalidRomanNumeralError(RomanError):
+    """
+    An exception to be raised when the input is not a valid Roman
+    numeral.
+    """
+
+    pass
+
+# ------------------------------------------------------------------------------
+#   Warnings
+# ------------------------------------------------------------------------------
 
 
 class PlasmaPyWarning(Warning):
@@ -69,6 +97,8 @@ class PhysicsWarning(PlasmaPyWarning):
 
     pass
 
+# ^^^^^^^^^^^^^ Base Warnings should be defined above this comment ^^^^^^^^^^^^^
+
 
 class RelativityWarning(PhysicsWarning):
     """
@@ -83,6 +113,8 @@ class CouplingWarning(PhysicsWarning):
     """
     A warning for functions that rely on a particular coupling regime to be valid.
     """
+
+    pass
 
 
 class ImplicitUnitConversionWarning(PlasmaPyWarning, UnitsWarning):
